Add explicit Door control class
